+++ lib/karafka/setup/configurators/water_drop.rb
@@ -7,8 +7,8 @@ module Karafka
         def setup
           ::WaterDrop.setup do |water_config|
             water_config.send_messages = true
-            water_config.connection_pool_size = config.concurrency
-            water_config.connection_pool_timeout = 1
+            water_config.connection_pool_size = config.connection_pool.size
+            water_config.connection_pool_timeout = config.connection_pool.timeout
             water_config.kafka.hosts = config.kafka.hosts
             water_config.raise_on_failure = true
           end